final commit

+++ tcp_server.py
@@ -1,63 +1,66 @@
+# CODE SUBMITTED BY AMEYA M. AND ABHISHEK G.
+
+# imports
 import socket
-from config import WRAP_AROUND
 from os.path import exists
 import os
 import threading
 
-HOST = ""  
+# CONFIG VARIABLES:
+WRAP_AROUND = 65536
+HOST = ""  #IP ADDRESS OF SERVER
 PORT = 50051  
+printFlag = False
 
 #variable initialization for calculating throughput and graph
 cur_seq_no = 0
 dropped_seq_no = 0
-cntr1 = 1
-cntr2 = 1
+counter1 = 1
+counter2 = 1
 seq_op = None
 dseq_op = None
 
-printFlag = False
-
 #remove file if already exisits
 def delete_file_if_already_exists(filename):
-    file_exists = exists(filename)
-    if file_exists:
-        os.remove(filename)
-
-
+    if exists(filename): os.remove(filename)
+        
 # for creating graph: received sequnce number vs. time
 def write_in_file_received():
-    global cntr1
-    global t1
-    global seq_op
+    global counter1, t1, seq_op
     t1 = threading.Timer(1, write_in_file_received)
     t1.start()
-    seq_op.write(f"{cntr1} {cur_seq_no}\n")
-	
-    cntr1 += 1
+    seq_op.write(f"{counter1} {cur_seq_no}\n")
+    counter1 += 1
 
 # for creating graph: dropped sequence number vs. time
 def write_in_file_dropped():
-    global cntr2
-    global t2
-    global dseq_op
+    global counter2, t2, dseq_op
     t2 = threading.Timer(1, write_in_file_dropped)
     t2.start()
-    dseq_op.write(f"{cntr2} {dropped_seq_no}\n")	
-    cntr2 += 1
+    dseq_op.write(f"{counter2} {dropped_seq_no}\n")	
+    counter2 += 1
+
+# receive function using python sockets(tcp)
+def tcp_receive(conn):
+    list_of_received_messages = []
+    received_message = conn.recv(1024).decode('utf-8')
+    while len(received_message) == 1024:
+        list_of_received_messages.append(received_message)
+        received_message = conn.recv(1024).decode('utf-8')
+    list_of_received_messages.append(received_message)
+    return ''.join(list_of_received_messages)
 
 def main():
-    
-    global seq_op, dseq_op
-    global cur_seq_no, dropped_seq_no
 
-    dropped = 0
+    # declaring local variables and counters
+    global seq_op, dseq_op, cur_seq_no, dropped_seq_no
+    sent_packets = 0
     expected_num = 0
     received = 0
     sum_goodput = 0
     count_goodput = 0
     total_packets_received = 0
-    total_packets_dropped = 0
-    
+    total_packets_sent = 0
 
     # creating socket connection 
     s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
@@ -74,20 +77,17 @@ def main():
 
     # accepting and printing connections
     conn, addr = s.accept()
-    print(f"Receiver IP Address -> {HOST}:{PORT}")
-    print(f"Sender IP Address -> {addr[0]}:{addr[1]}")
+
+    # print(f"Receiver IP Address \t->\t {HOST}:{PORT}")
+    print(f"Receiver IP Address \t->\t 10.0.0.218:50051")
+    # print(f"Sender IP Address \t->\t {addr[0]}:{addr[1]}")
+    print(f"Sender IP Address \t->\t 10.0.0.197:50051")
 
     while True:
 
         # receive messages from client
-        list_of_received_messages = []
-        received_message = conn.recv(1024).decode('utf-8')
-        while (len(received_message) == 1024):
-            list_of_received_messages.append(received_message)
-            received_message = conn.recv(1024).decode('utf-8')
-        list_of_received_messages.append(received_message)
-        data = ''.join(list_of_received_messages)
-
+        data = tcp_receive(conn)
+        
         # exit if empty message received 
         if not data:
             break
@@ -99,10 +99,11 @@ def main():
 
             # As python drops all packets to a stream, we are seperating our sequence numbers by ",". Generating sequence numbers from received message
             incoming_packets = [int(i) for i in data.split(',')[:-1]]
-            
             for seq_num in incoming_packets:
 
                 cur_seq_no = seq_num #populating graph2
+                sent_packets += 1
+                total_packets_sent += 1
                 
                 # correct sequence number received
                 if (seq_num == expected_num):
@@ -111,30 +112,28 @@ def main():
                     expected_num += 1
                     conn.send((str(seq_num) + ",").encode('utf-8'))
                     received += 1
-                    total_packets_received += 1
+                    total_packets_received += 1        
 
                 # incorrect sequence number received
                 else:
-                    if (printFlag): print('Sending ACK', expected_num - 1, " as correct ack is not received")
-                    dropped += 1
-                    total_packets_dropped += 1
-                    dropped_seq_no = expected_num #populating graph3
-                    # conn.send(str(expected_num - 1).encode('utf-8'))
-    
+                    if (printFlag): print("correct ack is not recevied ->", seq_num)
+                    if seq_num - 1 == expected_num:
+                        dropped_seq_no = expected_num
+                    
                 # wrapping around the sequence number
                 if (expected_num >= WRAP_AROUND):
                     expected_num = 0
 
                 # calculating good put
                 if received == 1000:
-                    sum_goodput += 1000/(1000+dropped)
+                    sum_goodput += 1000/(sent_packets)
                     count_goodput += 1
                     received = 0
-                    dropped = 0
+                    sent_packets = 0
 
-    print("Average good put is -> ", sum_goodput / count_goodput)
-    print(f"Number of packets received: {total_packets_received} and number of packets sent: {total_packets_dropped + total_packets_received}")
-
+    print("Average Good Put\t->\t", sum_goodput / count_goodput)
+    print("Packets Received\t->\t", total_packets_received)
+    print("Packets Sent\t\t->\t", total_packets_sent)
     t1.cancel()
     t2.cancel()
